+++ activerecord/lib/active_record/relation/query_methods.rb
@@ -303,6 +303,151 @@ module ActiveRecord
       self
     end
 
+    # Add a Common Table Expression (CTE) that you can then reference within another SELECT statement.
+    #
+    #   Post.with(:posts_with_tags, Post.where("tags_count > ?", 0))
+    #   # WITH posts_with_tags AS (
+    #   #   SELECT * FROM posts WHERE (tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts
+    #
+    # Once you define Common Table Expression you can use custom `FROM` value or `JOIN` to reference it.
+    #
+    #   Post.with(:posts_with_tags, Post.where("tags_count > ?", 0).from("posts_with_tags AS posts")
+    #   # WITH posts_with_tags AS (
+    #   #  SELECT * FROM posts WHERE (tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts_with_tags AS posts
+    #
+    #   Post.with(:posts_with_tags, Post.where("tags_count > ?", 0).joins("JOIN posts_with_tags ON posts_with_tags.id = posts.id"))
+    #   # WITH posts_with_tags AS (
+    #   #   SELECT * FROM posts WHERE (tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts JOIN posts_with_tags ON posts_with_tags.id = posts.id
+    #
+    # You can also pass the expression as a `String`.
+    #
+    #   Post.with(:posts_with_tags, "SELECT * FROM posts WHERE tags_count > 0")
+    #   # WITH posts_with_tags AS (
+    #   #   SELECT * FROM posts WHERE tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts
+    #
+    # Note that building your own string from user input may expose your application to injection attacks
+    # if not done properly. It is recommended to use `ActiveRecord::Relation` instead.
+    #
+    # To add multiple CTEs you can chain `.with` / `.with_recursive` calls
+    #
+    #   Post
+    #     .with(:posts_with_comments, Post.where("comments_count > ?", 0))
+    #     .with(:posts_with_tags, Post.where("tags_count > ?", 0))
+    #   # WITH posts_with_comments AS (
+    #   #   SELECT * FROM posts WHERE (comments_count > 0)
+    #   # ), posts_with_tags AS (
+    #   #   SELECT * FROM posts WHERE (tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts
+    #
+    # or pass them as a `Hash`.
+    #
+    #   Post.with(posts_with_comments: Post.where("comments_count > ?", 0), posts_with_tags: Post.where("tags_count > ?", 0))
+    #   # WITH posts_with_comments AS (
+    #   #   SELECT * FROM posts WHERE (comments_count > 0)
+    #   # ), posts_with_tags AS (
+    #   #   SELECT * FROM posts WHERE (tags_count > 0)
+    #   # )
+    #   # SELECT * FROM posts
+    #
+    # For recursive queries see `.with_recursive` method.
+    def with(name_or_opts, relation = nil)
+      opts = if relation
+        { name_or_opts => relation }
+      else
+        name_or_opts
+      end
+      spawn.with!(opts)
+    end
+
+    # Works same as `.with` but it adds `RECURSIVE` modifier to the query.
+    # Using RECURSIVE, a WITH query can refer to its own output.
+    #
+    #   non_recursive_query = Comment.select(:id, :parent_id).where(parent_id: nil)
+    #   recursive_query = Comment.select(:id, :parent_id).joins("JOIN threads ON threads.id = comments.parent_id")
+    #   Post.with_recursive(:threads, non_recursive_query, recursive_query)
+    #   # WITH RECURSIVE threads AS (
+    #   #   SELECT comments.id, comments.parent_id FROM comments WHERE comments.parent_id IS NULL
+    #   #   UNION ALL
+    #   #   SELECT comments.id, comments.parent_id FROM comments JOIN threads ON comments.parent_id = threads.id
+    #   # ) SELECT * FROM posts
+    #
+    # If you need to define columns with your common table name pass it as a `String`.
+    #
+    #   non_recursive_query = Comment.select(:id, :parent_id, 0).where(parent_id: nil)
+    #   recursive_query = Comment.select(:id, :parent_id, "threads.depth + 1").joins("JOIN threads ON threads.id = comments.parent_id")
+    #   Post.with_recursive("threads(id, parent, depth)", non_recursive_query, recursive_query)
+    #   # WITH RECURSIVE threads(id, parent, depth) AS (
+    #   #   SELECT comments.id, comments.parent_id, 0 FROM comments WHERE comments.parent_id IS NULL
+    #   #   UNION ALL
+    #   #   SELECT comments.id, comments.parent_id, threads.depth + 1 FROM comments JOIN threads ON comments.parent_id = threads.id
+    #   # ) SELECT * FROM posts
+    #
+    # Note that `UNION ALL` is used by default. If you need to use `UNION` you can set `union_all` option to `false`.
+    #
+    #   Post.with_recursive("threads(id, parent, depth)", non_recursive_query, recursive_query, union_all: false)
+    #   # WITH RECURSIVE threads(id, parent, depth) AS (
+    #   #   SELECT comments.id, comments.parent_id, 0 FROM comments WHERE comments.parent_id IS NULL
+    #   #   UNION
+    #   #   SELECT comments.id, comments.parent_id, threads.depth + 1 FROM comments JOIN threads ON comments.parent_id = threads.id
+    #   # ) SELECT * FROM posts
+    #
+    # To add multiple CTEs you can chain `.with` / `.with_recursive` calls
+    #
+    #   top_comment = Comment.order(replies_count: :desc).limit(1)
+    #   non_recursive_relation = Comment.select(:id, :parent_id, 0).joins("JOIN top_comment ON top_comment.id = comments.parent_id")
+    #   recursive_relation = Comment.select(:id, :parent_id, "replies.depth + 1").joins("JOIN replies ON comments.parent_id = replies.id")
+    #   Comment
+    #     .with(:top_comment, top_comment)
+    #     .with_recursive(:replies, non_recursive_relation, recursive_relation)
+    #   # WITH RECURSIVE top_comment AS (
+    #   #   SELECT comments.id FROM comments ORDER BY comments.replies_count DESC LIMIT 1
+    #   # ), replies(id, parent_id, depth) AS (
+    #   #   SELECT comments.id, comments.parent_id, 0 FROM comments JOIN top_comment ON top_comment.id = comments.parent_id
+    #   #   UNION ALL
+    #   #   SELECT comments.id, comments.parent_id, replies.depth + 1 FROM comments JOIN replies ON comments.parent_id = replies.id
+    #   # )
+    #   # SELECT comments.* FROM replies
+    #
+    # or pass them as a `Hash` but be aware that `ActiveRecord::Relation` does not have `UNION` support
+    # so you will probably need to pass the expression as `String` or raw Arel expression.
+    #
+    #   union = non_recursive_relation.arel.union(recursive_relation.arel)
+    #   Comment.with_recursive(top_comment: top_comment, replies: union)
+    #   # WITH RECURSIVE top_comment AS (
+    #   #   SELECT comments.id FROM comments ORDER BY comments.replies_count DESC LIMIT 1
+    #   # ), replies(id, parent_id, depth) AS (
+    #   #   SELECT comments.id, comments.parent_id, 0 FROM comments JOIN top_comment ON top_comment.id = comments.parent_id
+    #   #   UNION ALL
+    #   #   SELECT comments.id, comments.parent_id, replies.depth + 1 FROM comments JOIN replies ON comments.parent_id = replies.id
+    #   # )
+    #   # SELECT comments.* FROM replies
+    def with_recursive(name_or_opts, non_recursive_relation = nil, recursive_relation = nil, options = {})
+      self.recursive_with_value = true
+      if name_or_opts.is_a?(Hash)
+        opts = name_or_opts
+      else
+        options[:union_all] = true unless options.key?(:union_all)
+        union_args = options[:union_all] ? ["all", recursive_relation.arel] : recursive_relation.arel
+        union = non_recursive_relation.arel.union(*union_args)
+        opts = { name_or_opts => union }
+      end
+      with!(opts)
+    end
+
+    def with!(opts)
+      self.with_values += [opts]
+      self
+    end
+
     # Allows you to change a previously set select statement.
     #
     #   Post.select(:title, :body)
@@ -1266,6 +1411,7 @@ module ActiveRecord
         arel.group(*arel_columns(group_values.uniq)) unless group_values.empty?
 
         build_order(arel)
+        build_with(arel)
         build_select(arel)
 
         arel.optimizer_hints(*optimizer_hints_values) unless optimizer_hints_values.empty?
@@ -1406,6 +1552,44 @@ module ActiveRecord
           arel.project(*klass.column_names.map { |field| table[field] })
         else
           arel.project(table[Arel.star])
+        end
+      end
+
+      def build_with(arel)
+        return if with_values.empty?
+
+        with_statements = with_values.map do |with_value|
+          case with_value
+          when Arel::Nodes::As, Arel::Nodes::TableAlias then with_value
+          when Array then build_with_value_from_array(with_value)
+          when Hash then build_with_value_from_hash(with_value)
+          else
+            raise ArgumentError, "Unsupported argument type: #{with_value} #{with_value.class}"
+          end
+        end
+
+        recursive_with_value ? arel.with(:recursive, with_statements) : arel.with(with_statements)
+      end
+
+      def build_with_value_from_array(array)
+        if array.detect { |item| ![Arel::Nodes::As, Arel::Nodes::TableAlias].include?(item.class) }
+          raise ArgumentError, "Unsupported argument type: #{array} #{array.class}"
+        end
+
+        array
+      end
+
+      def build_with_value_from_hash(hash)
+        hash.map do |name, value|
+          expression =
+            case value
+            when Arel::Nodes::SqlLiteral, String then Arel.sql("(#{value})")
+            when ActiveRecord::Relation then value.arel
+            when Arel::SelectManager, Arel::Nodes::Union, Arel::Nodes::UnionAll then value
+            else
+              raise ArgumentError, "Unsupported argument type: #{value} #{value.class}"
+            end
+          Arel::Nodes::TableAlias.new(expression, Arel.sql(name.to_s))
         end
       end