+++ activerecord/test/cases/relation/delete_all_test.rb
@@ -57,6 +57,8 @@ class DeleteAllTest < ActiveRecord::TestCase
     assert_raises(ActiveRecord::ActiveRecordError) { Author.distinct.delete_all }
     assert_raises(ActiveRecord::ActiveRecordError) { Author.group(:name).delete_all }
     assert_raises(ActiveRecord::ActiveRecordError) { Author.having("SUM(id) < 3").delete_all }
+    assert_raises(ActiveRecord::ActiveRecordError) { Author.with(limited: Author.limit(2)).delete_all }
+    assert_raises(ActiveRecord::ActiveRecordError) { Author.with_recursive(limited: Author.limit(2)).delete_all }
   end
 
   def test_delete_all_with_joins_and_where_part_is_hash